Rename test/pathfinder to test/shortestpath (#220)
* Combine PrimitiveIntHashMap tests
* Remove unused TransportJsConverter
