+++ src/test/java/shortestpath/PluginMessageTest.java
@@ -1,12 +1,12 @@
-package pathfinder;
+package shortestpath;
 
 import net.runelite.client.RuneLite;
 import net.runelite.client.externalplugins.ExternalPluginManager;
-import shortestpath.ShortestPathPlugin;
 
-public class ShortestPathPluginTest {
+public class PluginMessageTest {
     public static void main(String[] args) throws Exception {
         ExternalPluginManager.loadBuiltin(ShortestPathPlugin.class);
+        ExternalPluginManager.loadBuiltin(PluginMessageTestPlugin.class);
         RuneLite.main(args);
     }
 }
