+++ src/test/java/shortestpath/ShortestPathPluginTest.java
@@ -1,13 +1,11 @@
-package pathfinder;
+package shortestpath;
 
 import net.runelite.client.RuneLite;
 import net.runelite.client.externalplugins.ExternalPluginManager;
-import shortestpath.ShortestPathPlugin;
 
-public class PluginMessageTest {
+public class ShortestPathPluginTest {
     public static void main(String[] args) throws Exception {
         ExternalPluginManager.loadBuiltin(ShortestPathPlugin.class);
-        ExternalPluginManager.loadBuiltin(PluginMessageTestPlugin.class);
         RuneLite.main(args);
     }
 }
